* fix types in writecache prototype (should be unsigned long int)
git-svn-id: https://svn.upfronthosting.co.za/svn/upfront/MonitorTools/trunk@52 80044c55-6f29-0410-b7d5-9c2a68987f97

+++ cpu.c
@@ -58,7 +58,7 @@ unsigned long int cpuusage(pid_t pid){
     return utime+stime;
 }
 
-void writecache(const char *cachefile, int uptime, int usage){
+void writecache(const char *cachefile, unsigned long int uptime, unsigned long int usage){
     int fd = open(cachefile, O_WRONLY | O_CREAT, S_IRUSR | S_IWUSR);
     if(fd!=-1){
         char buf[1024];
