+++ cpu.h
@@ -7,7 +7,7 @@
 void splitbuf(char*, size_t, char);
 unsigned long int getuptime();
 unsigned long int cpuusage(pid_t);
-void writecache(const char*, int, int);
+void writecache(const char*, unsigned long int, unsigned long int);
 void readcache(const char*, unsigned long int*, unsigned long int*);
 
 #endif
